+++ DonosServer/Core/Entities/Enums.cs
