--- DonosServer/Core1/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Core
-{
-    public class Class1
-    {
-    }
-}
