+++ DonosServer/DonosServer.sln
@@ -5,9 +5,9 @@ VisualStudioVersion = 16.0.30711.63
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "API", "DonosServer\API.csproj", "{1767769C-843A-4A3D-8EC2-0CCC7461D47A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core1\Core.csproj", "{1BE30474-323F-4527-A169-F41768B3ADB2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "Core\Core.csproj", "{D66226E1-DB64-44FF-9F50-2FFE8B6CFCA6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "Infrastructure1\Infrastructure.csproj", "{B7E0EBB3-81F4-4C5A-891C-C68C9CE93C8E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Infrastructure", "Infrastructure\Infrastructure.csproj", "{7A9943DF-5D2A-4334-9BFB-3719066786D0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,14 +19,14 @@ Global
 		{1767769C-843A-4A3D-8EC2-0CCC7461D47A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1767769C-843A-4A3D-8EC2-0CCC7461D47A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1767769C-843A-4A3D-8EC2-0CCC7461D47A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1BE30474-323F-4527-A169-F41768B3ADB2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1BE30474-323F-4527-A169-F41768B3ADB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1BE30474-323F-4527-A169-F41768B3ADB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1BE30474-323F-4527-A169-F41768B3ADB2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B7E0EBB3-81F4-4C5A-891C-C68C9CE93C8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B7E0EBB3-81F4-4C5A-891C-C68C9CE93C8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B7E0EBB3-81F4-4C5A-891C-C68C9CE93C8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B7E0EBB3-81F4-4C5A-891C-C68C9CE93C8E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D66226E1-DB64-44FF-9F50-2FFE8B6CFCA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D66226E1-DB64-44FF-9F50-2FFE8B6CFCA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D66226E1-DB64-44FF-9F50-2FFE8B6CFCA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D66226E1-DB64-44FF-9F50-2FFE8B6CFCA6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7A9943DF-5D2A-4334-9BFB-3719066786D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7A9943DF-5D2A-4334-9BFB-3719066786D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7A9943DF-5D2A-4334-9BFB-3719066786D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7A9943DF-5D2A-4334-9BFB-3719066786D0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
